requirements.txt
Django==1.5.5
pyotp==1.3.1
twilio==3.6.4
httpagentparser==1.5.0
django-tables2==0.14.0
django-filter==0.7
